feat: remove interfaces in exchange for types, implement pagination component and ability to create a 'fakepost'
a fake post is a post that is created straight away after the user creates a post and gets added to the feed. this post uses the real post id of the newly created post in the backend, however takes the rest of its data locally from the frontend. this is so that the user can directly see the post they have just created without having to refresh the entire feed and fetch all posts again. quite a cool little solution.